Map K8S namespace to Quobyte Tenant
Driver should be deployed with
--use_k8s_namespace_as_tenant=true. Defaultly
this option is disabled and should be enabled
by user.

Tenant is not created by Quobyte CSI plugin
and must be present in the Quobyte Storgate
before being used.

If mapping is enabled, StorageClass.parameters.
quobyteTenant is ignored and PVC.namespace is
used as quobyteTenant.