Document quota restriction with shared volumes
If Shared volumes are enabled, quota cannot be
set via dynamic provisioning. In other words,
createQuota in storage is not required and ignored
if provided.

This is due to the fact the many PVs are treated
as directory of the shared volume, and setting
Quota from one of the PV on shared volume is illogical
in this case.

If storage admin needs to restrict shared volume
storage consumption, they must set the quota for
the shared volume/tenant via management API/Quobyte
webconsole.