Add flag to disable volume metrics
If customer does not use Prometheus metrics of the volume, let
them disable it. The metrics execute statVolume and that is bit
expensive and causes load on Quobyte registry (client cache
can be configured to cache response via TTL to reduce the affect).

At the moment, there are efforts on registry side to improve
the statVolume call path, but if customer does not use the metric,
probing the stats is not useful.

For the backward compatibility, the flag is defaults to enabled state.