Disable metrics via flag (#87)
* Add flag to disable volume metrics

If customer does not use Prometheus metrics of the volume, let
them disable it. The metrics execute statVolume and that is bit
expensive and causes load on Quobyte registry (client cache
can be configured to cache response via TTL to reduce the affect).

At the moment, there are efforts on registry side to improve
the statVolume call path, but if customer does not use the metric,
probing the stats is not useful.

For the backward compatibility, the flag is defaults to enabled state.

* Fix test template snapshot

We have added a new flag that disables the volume metrics,
therefore a new snapshot base for tests should be generated.