Check k8s nodes ready check
E2E test case do not run and hang if any of the nodes
is in NotReady status. Therefore, wait for nodes to be in Ready
status before continuing with test setup.

In tests, it is confirmed that NotReady nodes is usually moved
to Ready after a couple of mins of restart.