Improvements to test cluster setup (#91)
* Release version v2.1.2 by ./build release command

* Add non-intuitive grep result code comment

* Check k8s nodes ready check

E2E test case do not run and hang if any of the nodes
is in NotReady status. Therefore, wait for nodes to be in Ready
status before continuing with test setup.

In tests, it is confirmed that NotReady nodes is usually moved
to Ready after a couple of mins of restart.

* Color code release instructions

* Add additional k8s node

Remove one node with the thought that the node not ready
is caused by exhausted resources. That is not the case,
so adding node back.

* Add DNS add hint to test cluster setup