Cleanup and improve helm unittests (#94)
* Remove unused CSI Driver resources yaml

* Use moved and up-to-date helm-unittest repo fork

The workflow is using a now deprecated project. Point
to the recent fork and also do not bind it to specific
version of helm-unittest so that we catch breakages fast.

* Run helm unittest as part of build process

* Remove unsupported helm version flag

* Update driver unittest snapshot

* Do not update helm unittest snapshot during build

...only run tests. Update requires manual verification
so add instructions.

* Remove volume config from examples

For Quobyte 2.x clusters, we expect upgrade to 3.x.
Nevertheless, keeping configuration for any existing
2.x clusters out there. For Quobyte version  (csi driver
values.yaml quobyte.version config) 2 only configuration
is used. For other versions, volume config is ignored.